Removed compilation warning

+++ Tests/UnitTests/TestCore/InstrumentTest.h
@@ -52,9 +52,9 @@ TEST_F(InstrumentTest, BeamManipulation)
     const IAxis &axis0 = m_instrument.getDetectorAxis(0);
     const IAxis &axis1 = m_instrument.getDetectorAxis(1);
     EXPECT_EQ( axis0.getName(), BA::PHI_AXIS_NAME );
-    EXPECT_EQ( axis0.getSize(), 10);
+    EXPECT_EQ( axis0.getSize(), (size_t)10);
     EXPECT_STREQ( axis1.getName().c_str(), "theta_f" );
-    EXPECT_EQ( axis1.getSize(), 20);
+    EXPECT_EQ( axis1.getSize(), (size_t)20);
 
     m_instrument.setBeamIntensity(10);
     EXPECT_EQ( double(10), m_instrument.getIntensity());
